polish(agen): extract actor gen code to own package
- Gen method accepts a `pkg` string, this allows other projects to
reuse the generator to generate their own actor shims with valid import
paths.
